use provider to switch between views
