+++ lib/models/nav_item.dart
@@ -1,7 +1,9 @@
+import 'package:memory_lamp/screens/home.dart';
+
 import 'icon_label_pair.dart';
 
 class NavItem extends IconLabelPair {
-  String routeName;
-  NavItem({required this.routeName, required name, icon, image, param})
+  Views view;
+  NavItem({required this.view, required name, icon, image})
       : super(name: name, icon: icon, image: image);
 }
