+++ lib/views/notification.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:memory_lamp/widgets/ml_text.dart';
+
+class NotificationView extends StatelessWidget {
+  const NotificationView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      child: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.all(24),
+          child: Column(
+            children: [MLText("notification")],
+          ),
+        ),
+      ),
+    );
+  }
+}
