+++ lib/widgets/compound_widgets/ml_drawer.dart
@@ -1,41 +1,43 @@
 import 'package:flutter/material.dart';
 import 'package:memory_lamp/models/nav_item.dart';
+import 'package:memory_lamp/providers/view_provider.dart';
 import 'package:memory_lamp/screens/home.dart';
 import 'package:memory_lamp/widgets/ml_text.dart';
+import 'package:provider/provider.dart';
 
 class MLDrawer extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     List<NavItem> _drawerItems = [
-      NavItem(icon: Icons.home, name: "Home", routeName: HomeScreen.routeName),
-      NavItem(icon: Icons.person, name: "Profile", routeName: "profile"),
-      NavItem(
-        icon: Icons.book_sharp,
-        name: "Browse Books",
-        routeName: "profile",
-      ),
+      NavItem(icon: Icons.home, name: "Home", view: Views.home),
+      NavItem(icon: Icons.person, name: "Profile", view: Views.profile),
+      NavItem(icon: Icons.book_sharp, name: "Browse Books", view: Views.books),
       NavItem(
         icon: Icons.notifications,
         name: "Notification",
-        routeName: "profile",
+        view: Views.notification,
       ),
-      NavItem(icon: Icons.settings, name: "Settings", routeName: "profile"),
-      NavItem(icon: Icons.logout, name: "Logout", routeName: "profile"),
+      NavItem(icon: Icons.settings, name: "Settings", view: Views.settings),
+      NavItem(icon: Icons.logout, name: "Logout", view: Views.home),
     ];
 
     return Drawer(
-      child: ListView(
-        children: List.generate(
-          _drawerItems.length,
-          (index) => ListTile(
-            leading: Icon(_drawerItems[index].icon),
-            title: MLText(_drawerItems[index].name),
-            onTap: () => Navigator.pushReplacementNamed(
-              context,
-              _drawerItems[index].routeName,
+      child: Consumer<ViewProvider>(
+        builder: (BuildContext context, viewProvider, Widget? child) {
+          return ListView(
+            children: List.generate(
+              _drawerItems.length,
+              (index) => ListTile(
+                leading: Icon(_drawerItems[index].icon),
+                title: MLText(_drawerItems[index].name),
+                onTap: () {
+                  viewProvider.changeView(_drawerItems[index].view);
+                  Navigator.pop(context);
+                },
+              ),
             ),
-          ),
-        ),
+          );
+        },
       ),
     );
   }
